core/vision.py
# backend/vision.py
from PIL import Image
import base64
import io

# You can swap this out with an actual LLaVA server or CLIP-based model
def analyze_image(file_stream):
    try:
        image = Image.open(file_stream).convert("RGB")
        # Placeholder for real image model processing
        return "🖼️ Image received. Vision model processing will go here."
    except Exception as e:
        return f"❌ Vision Error: {e}"

